+++ allauth/socialaccount/providers/oauth/client.py
@@ -6,6 +6,7 @@ Inspired by:
 """
 
 import requests
+from urllib.parse import parse_qsl, urlparse
 
 from django.http import HttpResponseRedirect
 from django.utils.http import urlencode
@@ -13,7 +14,6 @@ from django.utils.translation import gettext as _
 
 from requests_oauthlib import OAuth1
 
-from allauth.compat import parse_qsl, urlparse
 from allauth.utils import build_absolute_uri, get_request_param