+++ allauth/socialaccount/providers/oauth2/client.py
@@ -1,8 +1,7 @@
 import requests
+from urllib.parse import parse_qsl
 
 from django.utils.http import urlencode
-
-from allauth.compat import parse_qsl
 
 
 class OAuth2Error(Exception):