+++ allauth/socialaccount/providers/openid/provider.py
@@ -1,7 +1,8 @@
+from urllib.parse import urlparse
+
 from django.urls import reverse
 from django.utils.http import urlencode
 
-from allauth.compat import urlparse
 from allauth.socialaccount.providers.base import Provider, ProviderAccount
 
 from .utils import (