+++ ChangeLog.rst
@@ -1,3 +1,12 @@
+0.41.0 (Unreleased)
+*******************
+
+Backwards incompatible changes
+------------------------------
+
+- Dropped Python 2 and Django 1 compatibility.
+
+
 0.40.0 (2019-08-29)
 *******************